requirements.txt
audiolm-pytorch==1.1.4
fairseq~=0.12.2
tensorboardX
huggingface-hub
sentencepiece
transformers~=4.31.0
encodec~=0.1.1

numpy~=1.25.2
scipy~=1.11.1
torch~=2.0.1
setuptools~=65.5.1
funcy~=2.0
tqdm~=4.65.0
packaging~=23.1
accelerate~=0.21.0
einops~=0.6.1
torchaudio~=2.0.2
requests~=2.31.0
ipython~=8.14.0
azure-cognitiveservices-speech